Reimplements timer update saga to use fork instead of race effect
Thought of this switch in the shower. I think it's conceptually simpler,
not making the update race against the possible pause. The updating can
be a smaller task, and waiting for cancellation can be a smaller task.
It should also perform better, without having to tear down and recreate
the `take` effect for every update of a timer.